Revert the logic for fastbootd in FastbootMonitor
Reduce contention on fastbootMonitor, with the change it
requires to enumerate all the devices and run a command on
each, which is way slower and could cause late updated
state, which in turn would release device in bad state
or mess up device a bit after being flashed.

We will have to re-think how to detect fastbootd continuously.

Test: unit tests
Bug: 156620327
Bug: 148109220
Change-Id: I68df6c4226ff82c26702fda2655285ce10313594