Handle parsing output of failed subtest assertions
Adapt the Python unit test result parser to handle output from failed
subTest() assertions. The feature was introduced in Python 3 and makes
it possible to continue the test after the first failure which provides
more information for debugging.

Note that while this change supports parsing the output of multiple
failed subtest assertions, only the output of the last failed assertion
in a test is output on the TradeFed console. Users can inspect the log
files to see the test process output which contains the failed subtests
assertions.

See
https://docs.python.org/3/library/unittest.html#distinguishing-test-iterations-using-subtests for more information.

Test: m tradefed tradefed-tests && ./tests/run_tradefed_tests.sh
Bug: 156700051
Change-Id: Id26780080a28e4e91a85765839b4c37b5243c9d5