Add debugging message and fix a bug after de-duplicating mainline parameters.
This change does:
1. Add debugging message while a test config is mis-configured
   mainline parameters.
2. Fix bug after de-duplicating the parameters.

One case is: if a test config specifies "a+c+b" and TEST_MAPPING
defines the module name as "a+c+b", however, the parameters will
become "a+b+c" after the de-duplication, which is wrong.

Solution: Add a check to raise an exception if the parameters are
not configured in the alphabetical order.

Bug: 158545331
Bug: 158548420
Test: unittest.
      presubmit.
Change-Id: I84b7ee0090984c5625716a800542093dc1eec3d9