Adds some specific functionality for Robolectric
This should now be essentially complete to get the Robolectric tests
running in continuous infra.  The only thing that I don't see being
permanent is the classpath override option, but it is so useful for
debugging that it might be worth keeping around long term.

Added unit tests, more to come in the future.

Bug: b/157635117
Test: Unit
Change-Id: Ibf24860b0bcff8969bf4d3999b168228c8b033d8