Fix bug to handle module args for parameterized mainilne modules.
Module args can not work for parameterized mainline modules, the root
cause is the passed-in arguments in setUpConfig can not match the module
arg, which is like:
hello_world_test[mod1.apk]:exclude-annotation:com.android.testutils.SkipPresubmit.

The solution is to pass a new parameter(paramId) in setUpConfig so that
it can match the parameterized mainline module name.

Bug: 160675087
Test: forrest.
      unittest.
Change-Id: If7055bb5a735a335cfebb8012792fcbf308b3d1a