Don't change a command line when injecting subprocess reporters.
The new method creates a patched version of a target TF config and
bundles it in an injected jar as a same name along with subprocess
reporter classes instead of modifying a command line. This makes a
child TF to load the patched version with subprocess reporting
options instead of the original one.

This feature is meant to support legacy xTS versions up to Q.

Bug: 129111645
Test: unit test, ATS integration test.
Change-Id: I37711805f58a65534e00c26e180722e022917e16