Allow IPv6 link-local address to be fastboot serial number.
The fastbootd over ethernet was implemented with IPv6 link-local
address. An IPv6 address may include many ':'. It will break the fuzzy
test. And because the IPv6 addresses in normal mode is randomized and
the address used in fastboot mode isn't, we cannot use the normal mode
IPv6 link-local address to conect the device which is fastboot mode.

This change generates fastboot mode IPv6 link-local address from eth0
MAC address and uses it to connect devies in fastboot mode. And it will
only be enabled for devices which use TCP adb.

BUG: 158040108
BUG: 159570863
BUG: 159557111
Test: Test on both USB fastboot device and TCP fastboot device which
      uses IPv6 link-local
Change-Id: Iea598ce716e682aed0fae74370949440e34bc931