Pass option to optimize InstallApexModuleTargetPreparer for mainline.
Since InstallApexModuleTargetPreparer is a module-level preparer for
test mapping mainline testing, there is no way to enable a new option
(skip-apex-teardown) introduced in the preparer programmingly. This
CL provides a way to enable the option in the preparer.

The option(skip-apex-teardown) will be set when "optimize-mainline-test"
(aosp/1385762) which was added recently to reorder the test modules.

Bug: 162833142
Test: unittests
      forrest.
Change-Id: I2cccb717055d22d83a1311ae0e8491ff013ed278