Let TradeFed manage local virtual devices' instance directories
The instance directory contain log files that should be reported after
an instance is deleted. In the case of multiple instances, the deleted
instance's directory may be immediately cleared and reused by another
acloud process.

This commit makes TradeFed pass --local-instance-dir to acloud. The
directory is managed by TradeFed. It is not reused for other instances.

Bug: 160942364
Test: unit tests
Change-Id: Idf172622adbdd56ccaa35a41e58ccf14caaa7a12