Skip installing module if that are not preloaded for mainline test mapping.
On Q device, some modules are not preloaded on the device, which will cause
the InstallApexModuleTargetPreparer to raise an exception.
Because mainline test mapping creates the preparer dynamically, it's not
feasible to pass via --mainline-installer:ignore-if-module-not-preloaded
This change provide a way to skip installing module if they are not
preloaded.

Bug: 170927801
Test: unittests.
Change-Id: I79160f2aa177402bbe86be1d0bd5eecf99071a39