Adds the last couple fixes for Robolectric
There were two remaining fixes that needed to be added for Robolectric
tests to work properly under the subprocess runner. First, the
subprocess runner needs to be started in the directory of the test being
executed or the Robolectric configuration chain doesn't work correctly.
Secondly, we needed to modify the auto-includes on the classpath so it
didn't include multiple versions of Robolectric at once.

Bug: b/157635117
Test: manual
Change-Id: Id60fb8b4be270a5e90f071a1689cccd7543fe128