Ensures a test class has some testable methods.
If the class is using the JUnit4 style @Runwith annotation, it still
needs some testable methods to be considered runnable.  Probably should
come up with some formal designations of what will be automatically
picked up for test authors so they know to not put @Test methods in
their base class.

Bug: b/174168302
Test: WIP
Change-Id: I44a98daaf5642ab7dfa26cf2b8c714a1e8ad0569