Adds coverage for filtering for IsolatedHostTest
Since there is a lingering issue with the filtering mechanism
specifically in the context of automatic retry, we're adding some more
coverage to that part of the runner to enforce things that should be
true about correctly functioning filters.

Also, changed the package path of the reference tests to ensure they are
disambiguated from the isolation classes.  Not 100% necessary, but
useful to signal that the they are not a proper part of the isolation
environment.

Bug: b/175787126
Test: unit tests
Change-Id: Icf6c9255f1a2e3bf720c0bdd6f8938a1327635fe