Fixes filtering flow for the isolation runner
This makes the isolation runner ignore the "no tests to run" error from
JUnit when it has a filter parameter set.  This is because sometimes we
are excluding entire classes from the run set which means iterations of
the isolation runner will have no tests to run.  There are also a couple
of other niche scenarios where this makes the most sense.

Fixes: b/175787126
Bug: b/175787126
Test: unit tests
Change-Id: I616dd625a7ca3606e8263481abd7fcf9a9a073e1