Add support for module parameter groups.
The first group will be "multi_user" which maps to "run_on_work_profile"
and "run_on_secondary_user"

The basic behaviour is that the parameters ("run_on_work_profile"
and "run_on_secondary_user") behave normally. They can be included in
AndroidTest.xml, forced, excluded, etc.

Additionally, when an AndroidTest.xml specifies "multi_user" as a param,
this is resolved as if it included both "run_on_work_profile" and
"run_on_secondary_user".

Either individual parameters or "multi_user" can be specified as
--exclude-module-parameters. Only the indiviudal parameters can be used
when using --include-filter or --exclude-filter.

The changes are that instead of resolving forcedParameter to a single
param, we resolve it to multiple and then loop through the results. We
also expand excluded params and also all params when building the params
list.

Currently there are no parameter groups so all parameters expand to a
singleton set of themselves.

Existing tests are passing - showing no regressions. The next CL will
include a parameter group plus tests to ensure it behaves as expected.

Test: tapas tradefed-all && m && tools/tradefederation/core/tradefed.sh run singleCommand host -n --class com.android.tradefed.testtype.suite.SuiteModuleLoaderTest
Test: tapas tradefed-all && m && tools/tradefederation/core/tradefed.sh run singleCommand host -n --class com.android.tradefed.testtype.suite.params.ModuleParametersHelperTest
Change-Id: I2cdef5e0ff30df7a6e8597d61245ae93479f6fb2