RustBinaryTest: cd to test dir
b/171710847 added the "data" flag to blueprint files so that data
files can be passed into tests. The data files are stored in a relative
path to the test, so if the tests are run using the whole path, they
can't find their data files. Running the tests from the right dir allows
tests to find their data files.

Bug: 171710847
Test: atest parking_lot_core_device_test_src_lib
Test: atest quiche_device_test_src_lib
Change-Id: I5f50f8af69034047219b85f6bf782e5740f82341