Make SuiteApkInstaller compatible with RunOnWorkProfile target preparer.
This offers a way for an earlier target preparer to set a property which
later preparers can read to act on the correct user.

It will only change behaviour in cases where the property is set.

This allows a Module Parameter to insert the RunOnWorkProfile target
preparer and ensure that the package is installed into the correct user
(which is not known before the RunOnWorkProfile preparer is ran, so
can't be written into the configuration)

Test: tapas tradefed-all && m && tools/tradefederation/core/tradefed.sh run singleCommand host -n --class com.android.tradefed.targetprep.TestAppInstallSetupTest
Bug: 167945074
Change-Id: Ibd81668453b2da9aea8c36ebe0f35d4d1195d694