Improve error reporting in the ART run-test TradeFed test trunner.
The `com.android.tradefed.testtype.ArtRunTest` test runner used to
report any test error as soon as it was detected. To improve error
reporting (and make diagnosis easier), refactor the error checking
logic to first collect all errors and then report them (in a single
place) at the end of the test.

Test: m tradefed tradefed-tests \
        && tools/tradefederation/core/tests/run_tradefed_tests.sh \
             --class com.android.tradefed.testtype.ArtRunTestTest
Test: atest --test-mapping art:all
Bug: 171177621
Bug: 147812905
Change-Id: I611d077f2769f350ea6ba4630084620d0539ccc5