Fix and simplify Checker error reporting w.r.t. genuine test failures.
In case of a genuine Checker test failure (i.e. because of a failed
Checker assertion), we used to take an incorrect exceptional code path
that would hide the actual test failure and prevent the logging of
control-flow graph (CFG) information produced by `dex2oat` (that could
be used for further investigation); fix and simplify this logic.

Test: m tradefed tradefed-tests \
        && tools/tradefederation/core/tests/run_tradefed_tests.sh \
             --class com.android.tradefed.testtype.ArtRunTestTest
Test: Run `atest 'art-run-test-*-checker-*'` and check error messages
Bug: 162408889
Change-Id: I9c78da6445426d3af4186131e435ddf8daa016bc