ATest: Add quotes to classpath TF_PATH in atest_tradefed.sh.
When running atest_tradefed.sh with a wildcard classpath,
(which happens when no deps are added to the classpath by the script)
the expansion of the wildcard results in an invalid classpath.

Adding quotes solves for this failure, furthermore, since TF_PATH is
expected to always contain something, quotes are acceptable here.

This also matches the existing tradefed.sh script which quotes its
classpath.

Test: atest hello_world_test
Change-Id: I841867167603f0923ec92a91f184ff4f99bd22e2