Fix exit status detection for old devices.
Old devices (I don't remember exactly when, but pre-N or so) did not
have an adbd that supports shell_v2, which is required for exit status
propagation from the command run on the device to the caller. As such,
ExecutableTargetTest doesn't work on older devices.

To fix this, modify the command that is run to append the exit status
to stdout, and then parse that before returning the result to the
caller. This is only done for devices that don't have native support
for this behavior.

Test: Tested passing and failing ExecutableTargetTest on both a Jelly
      Bean and R device.
Bug: None
Change-Id: I9c59801715251e321fbcbe5a3c01d92cb3cc3d1a