Introduce vsr-min-api-level for ShippingApiLevelModuleController
To keep the original behavior of ShippingApiLevelModuleController,
we must not check the vendor api level for the "min-api-level"
option. Instead introduce a new option "vsr-min-api-level" to check
the vendor api level in addition to the device shipping api level.

A test that has vendor dependency must replace the "min-api-level"
option of the ShippingApiLevelModuleController to the new option
"vsr-min-api-level".

Bug: 181284704
Test: tools/tradefederation/core/javatests/run_tradefed_tests.sh \
--class com.android.tradefed.testtype.suite.module.ShippingApiLevelModuleControllerTest

Change-Id: I7008f329ed3fe04f8fd86942edb90a0a6dad6a83