Metric utility for metric file post processor.
This utility will be used to store the test results, write metrics to a
file.

It also provides utility method for aggregating similar tests which
differs only by iteration number in the test name or having same test
name with in same test run in Microbenchmark test runner.

Refactored AggregatePostProcessor to use MetricUtility.

Bug: b/177067595, b/186040059

Test: MetricUtilTest.
Change-Id: I845e1693a191b1b33bff761feb726876ff68742a