Reuse on-device apexes when running ART chroot tests.
Use only the ART test apex from the build, and reuse whatever
apexes are on the device for the remaining dependencies.

The original intent was to keep the ART tests hermetic so that
if ART depends on simultaneous changes in other apexes, the
changes will get picked up. This setup also allowed us to
run the tests on very old devices (potentially several years old).

This seems less relevant for eng-prod tests since the devices
are likely to more up to date (in fact, they are likely to be
flashed with the whole system image of the tested build).
We can also trust the apex API boundary stability more now.

Motivation: The conscrypt apex is switching to compressed mode.
Rather than supporting it, this removes the dependency.
It also removes the need for separate make command before atest.

Bug: 196361021
Bug: 187416712
Test: atest ArtGtestsTargetChroot
Change-Id: I287c1aadff833147656353392e9700ed40016750